bill email and pdf

+++ src/main/java/com/hospital/ERP/Repository/BillRepo.java
@@ -3,6 +3,6 @@ package com.hospital.ERP.Repository;
 import com.hospital.ERP.Entity.Bill;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BillRepo extends JpaRepository<Bill, Integer> {
+public interface BillRepo extends JpaRepository<Bill, Long> {
     
 }
